+++ hugegraph-api/src/main/java/com/baidu/hugegraph/api/traversers/CrosspointsAPI.java
@@ -37,8 +37,8 @@ import com.baidu.hugegraph.api.graph.VertexAPI;
 import com.baidu.hugegraph.backend.id.Id;
 import com.baidu.hugegraph.core.GraphManager;
 import com.baidu.hugegraph.server.RestServer;
+import com.baidu.hugegraph.traversal.algorithm.HugeTraverser;
 import com.baidu.hugegraph.traversal.algorithm.PathsTraverser;
-import com.baidu.hugegraph.traversal.algorithm.SubGraphTraverser;
 import com.baidu.hugegraph.type.define.Directions;
 import com.baidu.hugegraph.util.Log;
 import com.codahale.metrics.annotation.Timed;
@@ -81,11 +81,9 @@ public class CrosspointsAPI extends API {
 
         HugeGraph g = graph(manager, graph);
         PathsTraverser traverser = new PathsTraverser(g);
-        SubGraphTraverser.PathSet paths = traverser.paths(sourceId, dir,
-                                                          targetId, dir,
-                                                          edgeLabel, depth,
-                                                          degree, capacity,
-                                                          limit);
+        HugeTraverser.PathSet paths = traverser.paths(sourceId, dir, targetId,
+                                                      dir, edgeLabel, depth,
+                                                      degree, capacity, limit);
         return manager.serializer(g).writePaths("crosspoints", paths, true);
     }
 }
