+++ hugegraph-core/src/main/java/com/baidu/hugegraph/traversal/algorithm/SubGraphTraverser.java
@@ -291,15 +291,4 @@ public class SubGraphTraverser extends HugeTraverser {
             return true;
         }
     }
-
-    public static class PathSet extends HashSet<Path> {
-
-        public Set<Id> vertices() {
-            Set<Id> vertices = new HashSet<>();
-            for (Path path : this) {
-                vertices.addAll(path.vertices());
-            }
-            return vertices;
-        }
-    }
 }
